+++ lib/src/api/event/player_view/picture_in_picture_exit_event.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'picture_in_picture_exit_event.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+PictureInPictureExitEvent _$PictureInPictureExitEventFromJson(
+        Map<String, dynamic> json) =>
+    PictureInPictureExitEvent(
+      timestamp: json['timestamp'] as int?,
+    );
+
+Map<String, dynamic> _$PictureInPictureExitEventToJson(
+        PictureInPictureExitEvent instance) =>
+    <String, dynamic>{
+      'timestamp': instance.timestamp,
+    };